move signals computation to top-level obfuscation package
