+++ internal/staticanalysis/obfuscation/compute_signals_test.go
@@ -1,10 +1,9 @@
-package detection
+package obfuscation
 
 import (
 	"strings"
 	"testing"
 
-	"github.com/ossf/package-analysis/internal/staticanalysis/obfuscation"
 	"github.com/ossf/package-analysis/internal/staticanalysis/obfuscation/stats"
 	"github.com/ossf/package-analysis/internal/staticanalysis/obfuscation/stringentropy"
 	"github.com/ossf/package-analysis/internal/utils"
@@ -30,7 +29,7 @@ func compareSummary(t *testing.T, name string, expected, actual stats.SampleStat
 	}
 }
 
-func testSignals(t *testing.T, signals obfuscation.Signals, stringLiterals, identifiers []string) {
+func testSignals(t *testing.T, signals Signals, stringLiterals, identifiers []string) {
 	expectedStringEntropySummary := symbolEntropySummary(stringLiterals)
 	expectedStringLengthSummary := symbolLengthSummary(stringLiterals)
 	expectedIdentifierEntropySummary := symbolEntropySummary(identifiers)
@@ -56,7 +55,7 @@ func testSignals(t *testing.T, signals obfuscation.Signals, stringLiterals, iden
 
 func TestBasicSignals(t *testing.T) {
 	jsSource := `var a = "hello"`
-	rawData, err := obfuscation.CollectData(jsParserPath, "", jsSource, true)
+	rawData, err := CollectData(jsParserPath, "", jsSource, true)
 	if err != nil {
 		t.Error(err)
 	} else {
@@ -79,7 +78,7 @@ func TestBasicSignals2(t *testing.T) {
 			}
 		}
 	`
-	rawData, err := obfuscation.CollectData(jsParserPath, "", jsSource, true)
+	rawData, err := CollectData(jsParserPath, "", jsSource, true)
 	if err != nil {
 		t.Error(err)
 	} else {
